+++ plugins/catalog-backend/src/service/request/parseEntityTransformParams.test.ts
@@ -24,6 +24,11 @@ describe('parseEntityTransformParams', () => {
     metadata: {
       name: 'n',
       tags: ['t1', 't2'],
+      annotations: {
+        'example.test/url-like-key': 'ul1',
+        'example.com/other-url-like-key': 'ul2',
+        'other-example.test/next-url-like-key': 'ul3',
+      },
     },
     spec: {
       type: 't',
@@ -61,7 +66,115 @@ describe('parseEntityTransformParams', () => {
       parseEntityTransformParams({ fields: 'kind,metadata.name' })!(entity),
     ).toEqual({ kind: 'k', metadata: { name: 'n' } });
     expect(parseEntityTransformParams({ fields: 'metadata' })!(entity)).toEqual(
-      { metadata: { name: 'n', tags: ['t1', 't2'] } },
+      {
+        metadata: {
+          name: 'n',
+          tags: ['t1', 't2'],
+          annotations: {
+            'example.test/url-like-key': 'ul1',
+            'example.com/other-url-like-key': 'ul2',
+            'other-example.test/next-url-like-key': 'ul3',
+          },
+        },
+      },
     );
+  });
+
+  it('supports dot notated fields properly', () => {
+    expect(
+      parseEntityTransformParams({
+        fields: 'kind,metadata.annotations.example.com/other-url-like-key',
+      })!(entity),
+    ).toEqual({
+      kind: 'k',
+      metadata: { annotations: { 'example.com/other-url-like-key': 'ul2' } },
+    });
+  });
+
+  it('supports nested dot notated fields properly', () => {
+    entity.spec = {
+      ...entity.spec,
+      'field-with.dot': 'fd1',
+      'other-field-with.dot-also': {
+        subItem: 'fd2.sub',
+        'subite.with/dot': 'fd2.sub.dot',
+      },
+      'third-field-with.dot-again': 'fd3',
+    };
+
+    expect(
+      parseEntityTransformParams({
+        fields: 'kind,spec.other-field-with.dot-also',
+      })!(entity),
+    ).toEqual({
+      kind: 'k',
+      spec: {
+        'other-field-with.dot-also': {
+          subItem: 'fd2.sub',
+          'subite.with/dot': 'fd2.sub.dot',
+        },
+      },
+    });
+    expect(
+      parseEntityTransformParams({
+        fields: 'kind,spec.other-field-with.dot-also.subite.with/dot',
+      })!(entity),
+    ).toEqual({
+      kind: 'k',
+      spec: {
+        'other-field-with.dot-also': {
+          'subite.with/dot': 'fd2.sub.dot',
+        },
+      },
+    });
+  });
+
+  it('does not return a sub key if an incorrect longer key is requested', () => {
+    entity.spec = {
+      ...entity.spec,
+      strValue: 'st1',
+      boolValue: true,
+      numValue: 4,
+      arrValue: [4, 5],
+      nullValue: null,
+      undefValue: undefined,
+      'field-with.dot': 'fd1',
+    };
+
+    expect(
+      parseEntityTransformParams({ fields: 'kind,spec.strValue.other' })!(
+        entity,
+      ),
+    ).toEqual({ kind: 'k' });
+    expect(
+      parseEntityTransformParams({ fields: 'kind,spec.boolValue.other' })!(
+        entity,
+      ),
+    ).toEqual({ kind: 'k' });
+    expect(
+      parseEntityTransformParams({ fields: 'kind,spec.numValue.other' })!(
+        entity,
+      ),
+    ).toEqual({ kind: 'k' });
+    expect(
+      parseEntityTransformParams({ fields: 'kind,spec.arrValue.other' })!(
+        entity,
+      ),
+    ).toEqual({ kind: 'k' });
+    expect(
+      parseEntityTransformParams({ fields: 'kind,spec.nullValue.other' })!(
+        entity,
+      ),
+    ).toEqual({ kind: 'k' });
+    expect(
+      parseEntityTransformParams({ fields: 'kind,spec.undefValue.other' })!(
+        entity,
+      ),
+    ).toEqual({ kind: 'k' });
+    expect(
+      parseEntityTransformParams({ fields: 'kind,spec.field-with.dot.other' })!(
+        entity,
+      ),
+    ).toEqual({ kind: 'k' });
   });
 });