+++ plugins/kubernetes-backend/src/service/KubernetesBuilder.ts
@@ -318,6 +318,7 @@ export class KubernetesBuilder {
           name: cd.name,
           dashboardUrl: cd.dashboardUrl,
           authProvider: cd.authProvider,
+          oidcTokenProvider: cd.oidcTokenProvider,
         })),
       });
     });