+++ plugins/techdocs-react/src/addons.tsx
@@ -31,13 +31,13 @@ export const TECHDOCS_ADDONS_KEY = 'techdocs.addons.addon.v1';
 
 /**
  * Marks the <TechDocsAddons> registry component.
- * @alpha
+ * @public
  */
 export const TECHDOCS_ADDONS_WRAPPER_KEY = 'techdocs.addons.wrapper.v1';
 
 /**
  * TechDocs Addon registry.
- * @alpha
+ * @public
  */
 export const TechDocsAddons: React.ComponentType = () => null;
 
@@ -49,7 +49,7 @@ const getDataKeyByName = (name: string) => {
 
 /**
  * Create a TechDocs addon.
- * @alpha
+ * @public
  */
 export function createTechDocsAddonExtension<TComponentProps>(
   options: TechDocsAddonOptions<TComponentProps>,
@@ -96,7 +96,7 @@ const getAllTechDocsAddonsData = (collection: ElementCollection) => {
 
 /**
  * hook to use addons in components
- * @alpha
+ * @public
  */
 export const useTechDocsAddons = () => {
   const node = useOutlet();