+++ packages/backend-common/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @backstage/backend-common
 
+## 0.13.3-next.2
+
+### Patch Changes
+
+- e0a6360b80: Added a `stream()` method to complement the `buffer()` method on `ReadUrlResponse`. A `ReadUrlResponseFactory` utility class is now also available, providing a simple, consistent way to provide a valid `ReadUrlResponse`.
+
+  This method, though optional for now, will be required on the responses of `UrlReader.readUrl()` implementations in a future release.
+
+- 4b811aafce: Implemented the `UrlReader.search()` method for Google Cloud Storage. Due to limitations in the underlying storage API, only prefix-based searches are supported right now (for example, `https://storage.cloud.google.com/your-bucket/some-path/*`).
+- Updated dependencies
+  - @backstage/cli-common@0.1.9-next.0
+  - @backstage/config@1.0.1-next.0
+  - @backstage/config-loader@1.1.1-next.1
+  - @backstage/integration@1.2.0-next.1
+
 ## 0.13.3-next.1
 
 ### Patch Changes