+++ packages/techdocs-cli-embedded-app/CHANGELOG.md
@@ -1,5 +1,22 @@
 # techdocs-cli-embedded-app
 
+## 0.2.70-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.9.4-next.1
+  - @backstage/plugin-techdocs@1.1.1-next.2
+  - @backstage/plugin-techdocs-react@0.1.1-next.2
+  - @backstage/config@1.0.1-next.0
+  - @backstage/plugin-catalog@1.2.0-next.2
+  - @backstage/cli@0.17.1-next.2
+  - @backstage/catalog-model@1.0.2-next.0
+  - @backstage/core-app-api@1.0.2-next.1
+  - @backstage/core-plugin-api@1.0.2-next.1
+  - @backstage/integration-react@1.1.0-next.2
+  - @backstage/test-utils@1.1.0-next.2
+
 ## 0.2.70-next.1
 
 ### Patch Changes