+++ packages/test-utils/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @backstage/test-utils
 
+## 1.1.0-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/config@1.0.1-next.0
+  - @backstage/core-app-api@1.0.2-next.1
+  - @backstage/core-plugin-api@1.0.2-next.1
+  - @backstage/plugin-permission-common@0.6.1-next.0
+  - @backstage/plugin-permission-react@0.4.1-next.1
+
 ## 1.1.0-next.1
 
 ### Minor Changes