+++ plugins/catalog-backend-module-msgraph/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @backstage/plugin-catalog-backend-module-msgraph
 
+## 0.3.2-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-catalog-backend@1.1.2-next.2
+  - @backstage/backend-tasks@0.3.1-next.1
+  - @backstage/config@1.0.1-next.0
+  - @backstage/catalog-model@1.0.2-next.0
+
 ## 0.3.2-next.0
 
 ### Patch Changes