+++ plugins/catalog-graph/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @backstage/plugin-catalog-graph
 
+## 0.2.17-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.9.4-next.1
+  - @backstage/plugin-catalog-react@1.1.0-next.2
+  - @backstage/catalog-model@1.0.2-next.0
+  - @backstage/core-plugin-api@1.0.2-next.1
+  - @backstage/catalog-client@1.0.2-next.0
+
 ## 0.2.17-next.1
 
 ### Patch Changes