+++ plugins/catalog-graphql/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @backstage/plugin-catalog-graphql
 
+## 0.3.9-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/config@1.0.1-next.0
+  - @backstage/catalog-model@1.0.2-next.0
+
 ## 0.3.8
 
 ### Patch Changes