+++ plugins/config-schema/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @backstage/plugin-config-schema
 
+## 0.1.28-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.9.4-next.1
+  - @backstage/config@1.0.1-next.0
+  - @backstage/core-plugin-api@1.0.2-next.1
+
 ## 0.1.28-next.0
 
 ### Patch Changes