+++ plugins/graphiql/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @backstage/plugin-graphiql
 
+## 0.2.37-next.1
+
+### Patch Changes
+
+- e1f42a0e2c: Updated dependency `@types/codemirror` to `^5.0.0`.
+- Updated dependencies
+  - @backstage/core-components@0.9.4-next.1
+  - @backstage/core-plugin-api@1.0.2-next.1
+
 ## 0.2.37-next.0
 
 ### Patch Changes