+++ plugins/scaffolder-backend/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @backstage/plugin-scaffolder-backend
 
+## 1.2.0-next.1
+
+### Minor Changes
+
+- f8baf7df44: Added the ability to reference the user in the `template.yaml` manifest
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.13.3-next.2
+  - @backstage/plugin-catalog-backend@1.1.2-next.2
+  - @backstage/plugin-scaffolder-common@1.1.0-next.0
+  - @backstage/config@1.0.1-next.0
+  - @backstage/catalog-model@1.0.2-next.0
+  - @backstage/integration@1.2.0-next.1
+  - @backstage/catalog-client@1.0.2-next.0
+
 ## 1.2.0-next.0
 
 ### Minor Changes